Remove otb-bin from package install list
The new base image is based on Ubuntu 22.04, which (unlike 20.04
and 22.10) doesn't include the OTB packages in its standard repositories.
Since nobody has yet needed OTB in the exploitation system, and since
it seems to make more sense to include it in the thematic processing
system, omitting the package is the most straightforward solution.