Replace Dockerfile with all-in-one scipy-notebook
This commit temporarily removes all AVL customization and replaces
the AVL Dockerfile with a concatenation of the Dockerfiles for
docker-stacks-foundation, base-notebook, minimal-notebook, and
scipy-notebook from the Docker Stacks repository. The plan is to
test and adjust this new all-in-one scipy-notebook base, then re-add
the AVL steps as part of the same Dockerfile.

This commit also adds the additional files which the various Docker
Stacks Dockerfiles copy into their images.